Adding documentation on the notebook
I added documentation on how to access the notebook and start a panel dashboard for users.
I also removed the emission and states plot that was taking to much time to plot and resulted in infinite loading.
The documentation/tutorial should be tested during the following weeks and updated if necessary.